Final version of the CSS.
There was a problem with the responsiveness of the website. Even though the layout was satisfactory, when the width of the viewport was less than 600px, there was an empty gap at the right of the content (the whole HTML element was shrinking). Additional media queries have been added to ensure the correct display of the elements.